Let autosetup.sh be run more than once

+++ utilities/droplet.py
@@ -51,14 +51,18 @@ write_files:
         #!/bin/bash
 
         cd ~
-        echo 'export PATH=~/.local/bin:~/go/bin:$PATH' > ~/.zshrc-local
-        mkdir github; mkdir ship
-        (cd github; git clone https://github.com/ismail/hacks.git; git clone https://github.com/ismail/config.git)
-        (cd github/config; ./setup.sh; cd ../hacks; ./setup.sh)
+
+        if [ ! -d github ]; then
+            mkdir github; mkdir ship
+            (cd github; git clone https://github.com/ismail/hacks.git; git clone https://github.com/ismail/config.git)
+            (cd github/config; ./setup.sh; cd ../hacks; ./setup.sh)
+        fi
+
+        [ -f ~/.zshrc-local ] && echo 'export PATH=~/.local/bin:~/go/bin:$PATH' > ~/.zshrc-local
+        [ ! -f bin/youtube-dl ] && curl -sL https://yt-dl.org/downloads/latest/youtube-dl -o bin/youtube-dl; chmod +x bin/youtube-dl
+        [ ! -f bin/rar ] && curl -sO https://rarlab.com/rar/rarlinux-x64-5.6.1.tar.gz; tar xf rarlinux-x64-5.6.1.tar.gz; mv rar/rar rar/unrar bin; rm -rf rar*
 
         go get -u github.com/ncw/rclone; strip go/bin/rclone
-        curl -sL https://yt-dl.org/downloads/latest/youtube-dl -o bin/youtube-dl; chmod +x bin/youtube-dl
-        curl -sO https://rarlab.com/rar/rarlinux-x64-5.6.1.tar.gz; tar xf rarlinux-x64-5.6.1.tar.gz; mv rar/rar rar/unrar bin; rm -rf rar*
         curl -sO https://johnvansickle.com/ffmpeg/builds/ffmpeg-git-amd64-static.tar.xz; tar xf ffmpeg-git-amd64-static.tar.xz; mv ffmpeg-git*/ffmpeg bin; rm -rf ffmpeg*
     owner: ismail:ismail
     path: /etc/autosetup.sh
